+++ src/components/pages/about/About.js
@@ -1,6 +1,5 @@
 import React from 'react'; 
 
-import '../pageStyles.css'
 
 export const About = () => <div className="about">About</div>
