+++ src/components/pages/projects/Projects.js
@@ -1,6 +1,5 @@
 import React from 'react'; 
 
-import '../pageStyles.css'
 
 export const Projects = () => <div className="about">Projects</div>
